+++ src/config.hpp
@@ -36,12 +36,14 @@
 #include <magic_enum/magic_enum.hpp>
 #include <memory>
 #include <string>
+#include <vector>
 #include <ztl/enum.hpp>
 #include <ztl/implicit_wrapper.hpp>
 #include <ztl/limits.hpp>
 
 #if CONFIG_IDF_TARGET_ESP32S3
 #  include <driver/gptimer.h>
+#  include <esp_wifi.h>
 #  include <hal/adc_types.h>
 #elif CONFIG_IDF_TARGET_LINUX
 #  define ADC_CHANNEL_1 1
@@ -556,15 +558,13 @@ namespace wifi {
 
 inline constexpr auto led_gpio_num{GPIO_NUM_47};
 
+#if CONFIG_IDF_TARGET_ESP32S3
+inline std::vector<wifi_ap_record_t> ap_records;
+#endif
 inline std::string mdns_str;
 inline std::string ip_str;
 inline std::array<uint8_t, 6uz> mac;
 inline std::string mac_str(2uz * 6uz + 5uz + sizeof('\n'), '\0');
-
-///
-inline struct ApRecordsQueue {
-  QueueHandle_t handle{};
-} ap_records_queue;
 
 ///
 inline struct Task {